change asserts
asserts are mainly used for testing but in practice, they swallow up all the gas used for the transaction if it evaluates to false. You don't really want to use when you're validating user input for contract functions. It's much safer to use require as it refunds gas to user should sufficient conditions not be met.